feat:完善rabc权限控制

+++ acimage_community/src/main/java/com/acimage/community/service/topic/Impl/TopicInfoWriteServiceImpl.java
@@ -1,23 +1,20 @@
 package com.acimage.community.service.topic.Impl;
 
-import cn.hutool.core.bean.BeanUtil;
 import cn.hutool.core.util.StrUtil;
-import com.acimage.common.global.consts.FileFormat;
+import com.acimage.common.global.consts.FileFormatConstants;
 import com.acimage.common.global.enums.ServiceType;
 import com.acimage.common.model.Index.TopicIndex;
 import com.acimage.common.model.domain.community.Topic;
 import com.acimage.common.model.domain.community.TopicHtml;
 import com.acimage.common.model.mq.dto.HashImagesUpdateDto;
-import com.acimage.common.result.Result;
 import com.acimage.common.utils.*;
 import com.acimage.common.utils.common.FileUtils;
 import com.acimage.common.utils.common.ListUtils;
 import com.acimage.common.utils.minio.MinioUtils;
 import com.acimage.community.global.consts.CoverImageConstants;
-import com.acimage.community.global.consts.StorePrefixConst;
+import com.acimage.community.global.consts.StorePrefixConstants;
 import com.acimage.community.listener.event.PublishTopicEvent;
 import com.acimage.community.model.request.TopicAddReq;
-import com.acimage.community.model.request.TopicAddReqBak2;
 import com.acimage.community.model.request.TopicModifyHtmlReq;
 import com.acimage.community.mq.producer.HashImageMqProducer;
 import com.acimage.community.mq.producer.RemoveTopicImagesMqProducer;
@@ -30,7 +27,6 @@ import com.acimage.community.service.tag.TagTopicWriteService;
 import com.acimage.community.service.topic.*;
 import com.acimage.common.global.context.UserContext;
 import com.acimage.common.exception.BusinessException;
-import com.acimage.feign.client.ImageClient;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
@@ -60,8 +56,6 @@ public class TopicInfoWriteServiceImpl implements TopicInfoWriteService {
     @Autowired
     TagQueryService tagQueryService;
     @Autowired
-    ImageClient imageClient;
-    @Autowired
     HashImageMqProducer hashImageMqProducer;
     @Autowired
     SyncEsMqProducer syncEsMqProducer;
@@ -82,90 +76,19 @@ public class TopicInfoWriteServiceImpl implements TopicInfoWriteService {
 
 
     @Override
-    public long saveTopicAndImages(TopicAddReqBak2 topicAddReqBak2) {
-        //生成id
-        long topicId = IdGenerator.getSnowflakeNextId();
-        String serviceToken = topicAddReqBak2.getServiceToken();
-
-        //从image-service获取话题首张图url
-        String firstImageUrl;
-        Result<String> result = imageClient.updateTopicIdAndReturnFirstImageUrl(serviceToken, topicId);
-        if (result.isOk()) {
-            firstImageUrl = result.getData();
-        } else {
-            throw new BusinessException(result.getMsg());
-        }
-
-        Date now = new Date();
-        //转化
-        Topic topic = new Topic();
-        BeanUtil.copyProperties(topicAddReqBak2, topic, false);
-        topic.setId(topicId);
-        topic.setCreateTime(now);
-        topic.setUserId(UserContext.getUserId());
-        topic.setCoverImageUrl(firstImageUrl);
-        //插入topic
-        topicWriteService.save(topic);
-
-        //更新最新活跃时间
-        topicSpAttrWriteService.changeActivityTime(topicId, now);
-
-        //发送消息告诉image-service对图片哈希处理
-        hashImageMqProducer.sendHashImagesMessage(topicId);
-        //同步数据到es
-        PublishTopicEvent publishTopicEvent = new PublishTopicEvent(this, UserContext.getUserId(), topicId);
-        applicationContext.publishEvent(publishTopicEvent);
-        return topicId;
-    }
-
-    @Override
-    public long saveTopicAndImagesBak(TopicAddReqBak2 topicAddReqBak2) {
-        //生成id
-        long topicId = IdGenerator.getSnowflakeNextId();
-        String serviceToken = topicAddReqBak2.getServiceToken();
-
-        //从image-service获取话题首张图url
-        String firstImageUrl;
-        Result<String> result = imageClient.updateTopicIdAndReturnFirstImageUrl(serviceToken, topicId);
-        if (result.isOk()) {
-            firstImageUrl = result.getData();
-        } else {
-            throw new BusinessException(result.getMsg());
-        }
-
-        Date now = new Date();
-        //转化
-        Topic topic = new Topic();
-        BeanUtil.copyProperties(topicAddReqBak2, topic, false);
-        topic.setId(topicId);
-        topic.setCreateTime(now);
-        topic.setUpdateTime(now);
-        topic.setUserId(UserContext.getUserId());
-        topic.setCoverImageUrl(firstImageUrl);
-        //插入topic
-        topicWriteService.save(topic);
-
-        //更新最新活跃时间
-        topicSpAttrWriteService.changeActivityTime(topicId, now);
-        //发送消息告诉image-service对图片哈希处理
-        hashImageMqProducer.sendHashImagesMessage(topicId);
-        return topicId;
-    }
-
-    @Override
     public long saveTopicInfo(TopicAddReq topicAddReq, MultipartFile coverImage) {
         //生成id
         long topicId = IdGenerator.getSnowflakeNextId();
         Date now = new Date();
         String suffix = String.format("%s.%s", topicId, FileUtils.formatOf(coverImage));
-        String url = minioUtils.generateUrl(StorePrefixConst.COVER_IMAGE, now, suffix);
+        String url = minioUtils.generateUrl(StorePrefixConstants.COVER_IMAGE, now, suffix);
         //图片压缩
         InputStream inputStream = ImageUtils.compressAsFixedWebpImage(coverImage,
                 CoverImageConstants.WIDTH,
                 CoverImageConstants.HEIGHT,
                 CoverImageConstants.SIZE_AFTER_COMPRESS);
 
-        String coverImageUrl = minioUtils.upload(inputStream, url, FileFormat.WEBP_CONTENT_TYPE);
+        String coverImageUrl = minioUtils.upload(inputStream, url, FileFormatConstants.WEBP_CONTENT_TYPE);
         //过滤标题
         String filterTile = SensitiveWordUtils.filter(topicAddReq.getTitle());
         //过滤内容
@@ -195,13 +118,12 @@ public class TopicInfoWriteServiceImpl implements TopicInfoWriteService {
         //保存话题html
         topicHtmlWriteService.save(topicId, topicAddReq.getHtml());
         //保存标签
-
         tagTopicWriteService.save(topicId, noRepeatTagIds);
 
         //更新最新活跃时间
         topicSpAttrWriteService.changeActivityTime(topicId, now);
         //获取话题内的站内图片链接
-        List<String> newImageUrlList = HtmlUtils.getInnerImageUrlsAndRemoveRepeat(topicAddReq.getHtml());
+        List<String> newImageUrlList = HtmlUtils.getInnerImageUrls(topicAddReq.getHtml());
         HashImagesUpdateDto updateDto = HashImagesUpdateDto.builder()
                 .addImageUrls(newImageUrlList)
                 .topicId(topicId)
@@ -237,7 +159,41 @@ public class TopicInfoWriteServiceImpl implements TopicInfoWriteService {
             log.error("非法删除：用户非话题主人 话题:{} 用户:{}", topicId, UserContext.getUsername());
             throw new BusinessException("非法操作！话题不属于你");
         }
+        this.removeTopicInfoWithoutVerification(topicId);
 
+//        //删除star
+//        starWriteService.removeStars(topicId);
+//        //删除评论
+//        commentWriteService.removeComments(topicId);
+//        //删除话题
+//        topicHtmlWriteService.remove(topicId);
+//        topicWriteService.remove(topicId);
+//        //删除标签
+//        tagTopicWriteService.remove(topicId);
+//        //删除相关属性
+//        topicSpAttrWriteService.removeAttributes(topicId);
+//        //更新用户统计数据
+//        cmtyUserWriteService.updateTopicCountByIncrement(userId, -1);
+//        //发送删除图片的消息
+//        removeTopicImagesMqProducer.sendRemoveTopicMessage(topicId);
+//        //同步es数据
+//        syncEsMqProducer.sendDeleteMessage(Long.toString(topicId), TopicIndex.class);
+//        //同步到以图识图
+//        HashImagesUpdateDto updateDto = HashImagesUpdateDto.builder()
+//                .topicId(topicId)
+//                .serviceType(ServiceType.DELETE)
+//                .build();
+//        hashImageMqProducer.sendHashImagesMessage(updateDto);
+    }
+
+    @Override
+    public void removeTopicInfoWithoutVerification(long topicId){
+        Topic topic = topicQueryService.getTopic(topicId);
+        if (topic == null) {
+            log.error("话题不存在 话题:{} 用户:{}", topicId, UserContext.getUsername());
+            throw new BusinessException("话题不存在~~");
+        }
+        long userId=topic.getUserId();
         //删除star
         starWriteService.removeStars(topicId);
         //删除评论
@@ -250,7 +206,7 @@ public class TopicInfoWriteServiceImpl implements TopicInfoWriteService {
         //删除相关属性
         topicSpAttrWriteService.removeAttributes(topicId);
         //更新用户统计数据
-        cmtyUserWriteService.updateTopicCountByIncrement(topic.getUserId(), -1);
+        cmtyUserWriteService.updateTopicCountByIncrement(userId, -1);
         //发送删除图片的消息
         removeTopicImagesMqProducer.sendRemoveTopicMessage(topicId);
         //同步es数据
@@ -287,8 +243,8 @@ public class TopicInfoWriteServiceImpl implements TopicInfoWriteService {
         HashImagesUpdateDto updateDto = null;
         TopicHtml topicHtml = topicHtmlQueryService.getTopicHtml(topicId);
         if (topicHtml != null) {
-            List<String> oldUrls = HtmlUtils.getInnerImageUrlsAndRemoveRepeat(topicHtml.getHtml());
-            List<String> newUrls = HtmlUtils.getInnerImageUrlsAndRemoveRepeat(modifyReq.getHtml());
+            List<String> oldUrls = HtmlUtils.getInnerImageUrls(topicHtml.getHtml());
+            List<String> newUrls = HtmlUtils.getInnerImageUrls(modifyReq.getHtml());
             List<String> addImageUrls = ListUtils.differenceSetOfV2(newUrls, oldUrls);
             List<String> removeImageUrls = ListUtils.differenceSetOfV2(oldUrls, newUrls);
             updateDto = HashImagesUpdateDto.builder()
